+++ src/main/java/com/besson/arknights/block/ModBlockEntities.java
@@ -3,6 +3,9 @@ package com.besson.arknights.block;
 import com.besson.arknights.ArknightsFurniture;
 import com.besson.arknights.block.warehouse.CartonBE;
 import com.besson.arknights.block.warehouse.LargeShelfBE;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.MobCategory;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
@@ -17,6 +20,13 @@ public class ModBlockEntities {
     public static final DeferredRegister<BlockEntityType<?>> BLOCK_ENTITIES =
             DeferredRegister.create(ForgeRegistries.BLOCK_ENTITY_TYPES, ArknightsFurniture.MOD_ID);
 
+    public static final DeferredRegister<EntityType<?>> ENTITY_TYPES =
+            DeferredRegister.create(ForgeRegistries.ENTITY_TYPES, ArknightsFurniture.MOD_ID);
+
+    public static final RegistryObject<EntityType<SeatEntity>> SEAT =
+            register("seat", EntityType.Builder.<SeatEntity>of((type, world) ->
+                    new SeatEntity(world), MobCategory.MISC).sized(0.0F, 0.0F).setCustomClientFactory((spawnEntity, world) -> new SeatEntity(world)));
+
     public static final RegistryObject<BlockEntityType<CartonBE>> CARTON =
             register("carton", CartonBE::new,
                     () -> new Block[]{ModBlocks.CARTON.get()});
@@ -29,6 +39,11 @@ public class ModBlockEntities {
                             ModBlocks.LARGE_SHELF3.get(),
                             ModBlocks.LARGE_SHELF4.get()});
 
+    private static <T extends Entity> RegistryObject<EntityType<T>> register(String name, EntityType.Builder<T> builder)
+    {
+        return ENTITY_TYPES.register(name, () -> builder.build(name));
+    }
+
     private static <T extends BlockEntity> RegistryObject<BlockEntityType<T>> register(
             String name, BlockEntityType.BlockEntitySupplier<T> supplier, Supplier<Block[]> validBlocksSupplier) {
         return BLOCK_ENTITIES.register(name,
@@ -37,5 +52,6 @@ public class ModBlockEntities {
 
     public static void register(IEventBus eventBus){
         BLOCK_ENTITIES.register(eventBus);
+        ENTITY_TYPES.register(eventBus);
     }
 }