+++ src/main/java/com/besson/arknights/block/SofaBlock.java
@@ -1,7 +1,23 @@
 package com.besson.arknights.block;
 
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
+
 public class SofaBlock extends ConnectBlock{
     public SofaBlock(Properties pProperties) {
         super(pProperties);
+    }
+
+    @Override
+    public InteractionResult use(BlockState pState, Level pLevel, BlockPos pPos, Player pPlayer, InteractionHand pHand, BlockHitResult pHit) {
+        if (!pLevel.isClientSide()){
+            return SeatEntity.create(pLevel, pPos, 0.25, pPlayer, pState.getValue(FACING));
+        }
+        return InteractionResult.PASS;
     }
 }