+++ src/main/java/com/besson/arknights/block/warehouse/CartonStool.java
@@ -1,10 +1,16 @@
 package com.besson.arknights.block.warehouse;
 
 import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.SeatEntity;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 
@@ -18,5 +24,13 @@ public class CartonStool extends FurnitureHorizontalFacingBlock {
     @Override
     public VoxelShape getShape(BlockState pState, BlockGetter pLevel, BlockPos pPos, CollisionContext pContext) {
         return SHAPE;
+    }
+
+    @Override
+    public InteractionResult use(BlockState pState, Level pLevel, BlockPos pPos, Player pPlayer, InteractionHand pHand, BlockHitResult pHit) {
+        if (!pLevel.isClientSide()) {
+            return SeatEntity.create(pLevel, pPos, 0.65, pPlayer, pState.getValue(FACING));
+        }
+        return InteractionResult.PASS;
     }
 }