+++ src/main/java/com/besson/arknights/renderer/SeatEntityRenderer.java
@@ -1,4 +1,25 @@
 package com.besson.arknights.renderer;
 
-public class SeatEntityRenderer {
+import com.besson.arknights.block.SeatEntity;
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.entity.EntityRenderer;
+import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+
+public class SeatEntityRenderer extends EntityRenderer<SeatEntity> {
+    public SeatEntityRenderer(EntityRendererProvider.Context pContext) {
+        super(pContext);
+    }
+
+    @Override
+    public ResourceLocation getTextureLocation(SeatEntity pEntity) {
+        return null;
+    }
+
+    @Override
+    protected void renderNameTag(SeatEntity pEntity, Component pDisplayName, PoseStack pPoseStack, MultiBufferSource pBuffer, int pPackedLight) {
+
+    }
 }